Added timeseries and aggregations, and also added feeCollector info to strategies
